explicitly cast to int and float

+++ src/mqtt/MQTT.cpp
@@ -693,9 +693,9 @@ std::string MQTT::meshPacketToJson(meshtastic_MeshPacket *mp)
     jsonObj["type"] = new JSONValue(msgType.c_str());
     jsonObj["sender"] = new JSONValue(owner.id);
     if (mp->rx_rssi != 0)
-        jsonObj["rssi"] = new JSONValue(mp->rx_rssi);
+        jsonObj["rssi"] = new JSONValue((int)mp->rx_rssi);
     if (mp->rx_snr != 0)
-        jsonObj["snr"] = new JSONValue(mp->rx_snr);
+        jsonObj["snr"] = new JSONValue((float)mp->rx_snr);
 
     // serialize and write it to the stream
     JSONValue *value = new JSONValue(jsonObj);
